add import_object_ns function
To support bp:virt-driver-cleanup, there is a desire to have shorter
driver strings (i.e. libvirt.LibvirtDriver instead of
nova.virt.libvirt.LibvirtDriver). One way to support this is with
a new import_object variant that takes a namespace to load from.

If the code fails to load the driver in the prefered namespace it
will also try loading the full driver path, before throwing an
exception.

Change-Id: Ib97c92f38685ca89f29890f8015fc413acc744f8