Import timeutils.is_soon from keystoneclient
This is a function defined as will_expire_soon at least twice in
python-keystoneclient, and it's actually quite handy. There is a comment on
that code about moving to timeutils, and it actually sounds like a good idea
since I'm likely to use it in Ceilometer, so here it is.

Change-Id: Idf208ce9e7ce9048a38d98212b7f730c6b7c8288
Signed-off-by: Julien Danjou <julien@danjou.info>