urlsplit issues with IPv6 addresses in python26
urlsplit can't handle RFC 2732 URLs. reuse the
parse_host_port to rectify the hostname and
port implementation of SplitResult. Added test
cases to make sure we get the right ip and port
when the urls have ipv6 in them. Added an additional
test for both ipv4 and ipv6 to ensure we are indeed
handling username password in urls as well

Change-Id: I054f5e46cc8c201aaa48ee2d8b9d0a09b30a48ce
Closes-Bug: #1298137