Provide a common exception caused by base class
Both taskflow and tooz (and others of `raise_with_cause`
should be able to use this as there base exception class
to ensure that there usage of `raise_with_cause` works
as expected).

This adds the class that taskflow was/is using that provides
a 'good enough' version exception chaining (and a decently useful
`pformat` and `__str__` overrides that approximates the exception
chaining message on py2.x).

When used correctly this produces the following:

http://paste.openstack.org/show/326021/

Which is a 'good enough' approximation of exception chaining
on py2.x (and just uses the native exception chaining on py3.x)

This also adds some basic tests to ensure the functionality
provided continues to work as expected.

Change-Id: If814816dec91f7da3c90f4b9b8e2d3cf7228b928