improve tests for mask_password and mask_dict_password
All the tests used a very 'strong' password of mypassword and it
wasn't apparent in reviewing some new tests whether the masking was
being done of the 'word' password in mypassword or the command option
or dictionary key.

So I've taken the liberty to rewrite all the tests with a better
password.

Change-Id: I50c05d79bc9d8eaae5147d7f96ef5abfd0f2b2e1