Restore <all-in> operator
We removed the <all-in> operator as it was blocking much-needed
refactoring of the specs matcher and thereby any release of
oslo.utils. This resolves the interface question using
ast.literal_eval, that is to say a string must always be passed but if
the expression uses <all-in>, that string must be a Python list
literal. As far as I can tell, this should be compatible with Nova's
usage.

Change-Id: Iac8b35af8362d5d1fc1b57ab2dc50df795287b45