Add utils for validating and splitting quotes
Now Murano and Glance have similiar implementations of validating
and splitting quotes for using operator 'in' in their API [1],[2].

The style of this functionality should be common as they follow
appropriate api-wg style [3]. So it would be great to have them
implemented in one place and reuse in different projects.

TBD: add appropriate commits to Murano and Glance if this patch
isn't revoked.

[1]: https://github.com/openstack/glance/blob/master/glance/common/utils.py#L601-L644
[2]: https://github.com/openstack/murano/blob/master/murano/common/utils.py#L250-L301
[3]: https://specs.openstack.org/openstack/api-wg/guidelines/pagination_filter_sort.html#filtering

Change-Id: I15236d28b2136153db093590fb4dae65a87eb3e3
Co-Authored-By: Alexis Lee <lxsli@hpe.com>