Create dictutils and add 'flatten_dict_to_keypairs'
This function has been shared at least kwapi and all Telemetry projects
including ceilometer, aodh, panko and gnocchi. As a generator, it produces
sequence of keypairs for nested dictionaries. It seems to be good moving it
to the common shared code.

So add an initial dictutils and tests for it.

Change-Id: Ibd7c2ce4ef9c2608910b26169a804eb463759c7f