Handle non-string keys appropriately
In python dict keys can be anything that is hashable; which
includes non-strings such as ints. Currently the code is blowing
up with these types of keys with exceptions like:

TypeError: argument of type 'int' is not iterable

So to fix that handle the case where non-string keys are found.

Change-Id: I4f576a089df6f68e43572bf0eee15e99f2b557fe