Avoid calling eventlet.event.Event.reset()
The eventlet maintainers have made it clear that calling reset() on an
Event is not recommended:

https://github.com/eventlet/eventlet/commit/715b2ced5239bbada84c06f488b9c8dda8180230

We don't really need to resend the event to wake up threads, because
Event.wait() will return immediately once the event has been sent. So
just rely on the internal _set flag to tell us whether we need to send
the event.

Change-Id: I466aa7cb64308e018598c3bb63a9d0cfbc833adc