Avoid double-setting event
In commit cc8b51e1e16f6bdc7d6c0e571e2002e70cde098d we added a send
on the old event when clearing an EventletEvent. However, this was
done unconditionally, which means if the event was already sent
then we attempt to send it again. This fails with:

AssertionError: Trying to re-send() an already-triggered event.

Similar to 14a53c4d8a9d5605b14a503803859df0d6d4b820, we should check
if self._set is True and if so then we know that the event was
already sent and we don't need to do it again.

Change-Id: I660601383072d11e4a077aada8c1b8c30b9d8d1d
Closes-Bug: 1812922