Dropping lower constraints testing
We facing errors related to the new pip resolver, this
topic was discussed on the ML and QA team proposed to
to test lower-constraints [1].

I propose to drop this test because the complexity and recurring pain needed
to maintain that now exceeds the benefits provided by this mechanismes.

Contain also the patch[2] to remove the tox environment and the patch[3] to
remove the lower-constraints.txt file.

https://review.opendev.org/c/openstack/oslo.utils/+/787867

[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-December/019390.html
[2] https://review.opendev.org/c/openstack/oslo.utils/+/772630
[3] https://review.opendev.org/c/openstack/oslo.utils/+/787867

Change-Id: Ia02154a1eb04434658a21adf82599ae55302e310
(cherry picked from commit c4e3576c63f8f12dc1b2c3ff9bf0350042cd1703)