Drop lower-constraints
Lower-constraints is not a requirement of the OpenStack Python PTI
[0] and there currently is a discussion on the mailing list [1]
about dropping the test, with the oslo team already having done
so [2].

The new dependency resolver in pip fails due to incompatible
dependency versions in our lower-constraints file, meaning that
we were never providing any real guarantees with it.

To unblock the CI, I am disabling lower-constraints job for now,
with the option to reenable it in case we fix the constraints,
and based on the outcome of the mailing list discussions and
consensus.

[0]. https://governance.openstack.org/tc/reference/pti/python.html
[1].
http://lists.openstack.org/pipermail/openstack-discuss/2021-January/019672.html
[2].
http://lists.openstack.org/pipermail/openstack-discuss/2021-January/019659.html

Change-Id: I8c647b325a6f489aaffdab474117c5451852c81e