QemuImgInfo: Skip deprecation warning when output is not passed
Change 73eb0673f627aad382e08a816191b637af436465 deprecated usage of
the human format and introduced a deprecatipon warning message which
is shown unless format is explicitly set to 'json'.

To avoid the deprecation warning, all usage of QemuImgInfo requires
explicit definition of format='json'. This means that we should add
the format parameter to all existing usage of QemuImgInfo even if
output is blank as is described in the following example.
 QemuImgInfo()
 QemuImgInfo(output=None)

However later we should revert these implementations again when we
deprecated or remove the format parameter. These steps are very
redundant.

This change suppresses the warning message when output is blank so that
we can avoid deprecation warnings without redundant update and revert.

Change-Id: If1ec42dae757fa3d74c740a52c346701ea19f1c9