Make imageutils a directory
In anticipation of more image utility functionality incoming (i.e.
format_inspector), this makes imageutils a directory so we can put
that stuff underneath. The original imageutils.py file only has the
QemuImgInfo object in it, so that is moved to imageutils/qemu.py and
is imported in the module in the same place so this does not change
the API.

Change-Id: I3a58312d3bff69d9b06b712af466e6e09bf0b575