Import format_inspector from nova
This adds the format_inspector module to imageutils/. This started in
glance, but was brought into nova and cinder in order to address a CVE,
with the ultimate intention of it moving to oslo as soon as possible.

Glance is technically the original and reference implementation, but
Nova extended it with an ISO9660 format inspector shortly after merge
to handle a functional regression that was introduced. Thus, this is
Nova's copy although it's equivalent to Glance's plus that addition.

There are things we need to do before we commit to an interface, but
I wanted to bring it in with minimal changes as a starting point so
that those deltas are more easily identified.

The changes from nova to oslo-ify it are:
- Removal of the asserts which fail the pep8 check, and were only for
  sanity checking during development of the ISO support. These limits
  are effectively checked by the struct module when parsing and by
  the membership check for the signature.
- Use of logging instead of oslo_log
- Use of oslo's own BaseTestCase

Change-Id: I1c6b59ee1dfde29a4e3ded9feab9932718929d85