Make FileInspector an ABC
This is in response to some feedback about the module during review.
It makes FileInspector an ABC with some required implementations by
subclasses and also makes 'raw' a dedicated inspector instead of just
being the base class for everything else. The fact that it includes
no test change shows that it doesn't alter the runtime behavior at
all. The added test is just for the simplification of unifying the
__str__ implementation so that it doesn't need to be implemented by
every subclass.

Change-Id: I4cedbaf4f2128f89b2e18f2f84c5df8c9550dd6e