Modularize image safety checks
The original implementation made the safety check an all-or-nothing
operation for simplicity. That ended up not being suitable for (at
least) one use-case, which needed to allow one check that would
normally fail.

This change adds a SafetyCheck wrapper class and makes each format
define at least one (even if it is just a no-op). Those checks are
all run when safety_check() is executed and an exception is raised if
one or more of them fail. The checks are enumerated in the exception,
so that excluding one of them for an expected/allowable reason will
be doable by callers that have the need.

Co-Authored-By: Takashi Kajinami <kajinamit@oss.nttdata.com>
Change-Id: I4e0a4f727b99a11fc0fd831b805300e49d3bce9c