Add region-complete signaling for inspectors
As format_inspector has grown in utility, the redundancy of re-parsing
the region data each time has become excessive. We avoided refactoring
that infrastructure while addressing the CVE fix, but now is the time.
This adds a method that an inspector can implement to receive a
callback when a region becomes complete so it can parse it once. The
post_process() callback is similar, but it is called once per chunk,
instead of once per region. Parsing the region once improves the
performance for multiple calls to get information and safety-check
the content.

NOTE: This maybe fixes a VMDK parsing bug where if we get a mis-aligned
chunk, we may fail to unpad it because the region doesn't end with
a pad character and thus rindex() does not find it. Since the
descriptor region is variable-length, this could happen and/or could
be an attack vector for fooling our matching. We actually have a test
for this already but this change tickled it to fail in a different way.
Thus the fixed code is still confirmed by the test. Additional tests
are added for tightened checks.

Change-Id: I7caa96feae5394781be48476ed88e0db2549f668