Support VMDK sparse with footer
This allows us to capture the last N bytes of a stream and removes the
rejection of VMDKs with a footer.

Doing this requires quite a bit more infrastructure, which is added
here. The "capture last N bytes" requires a new region type that has
only an offset from EOF, and which has to be marked as complete
manually since we do not know when EOF is coming until it arrives. The
requirement to signal EOF also changes our detection logic a bit as
any format with an EndCaptureRegion will not signal match until EOF.
To confirm that this works in the presence and absence of such a
region, an additional check to make sure we detect our tested files
is added if/when we signal that the safety check should pass for a
given test case.

Allowing the VMDK footer requires additional checks of that footer,
ensuring it does not redefine anything from the header that we have
checked, and/or that we have missed our opportunity _to_ check, given
we can not go back in the stream after the fact. Since qemu-img will
just seek() to the footer and re-read it, we must make sure we do not
allow a footer than defines a different descriptor than the one we
validated earlier.

Also note that this corrects a few of the offsets in the VMDKInspector
docstring, which were off by four bytes and also adds missing ones
added after the footer check was added.

Related-Bug: #2073262
Change-Id: I96c0f7d49972fb45a3596f684be09f616a9b3672