Fix qcow2 feature flag checks (for the future)
This is not a problem now, but in a hypothetical future where qcow2
has grown more than double the features it has added since inception,
we need to be able to check feature bits >8. Due to an oversight in
the feature flag mask calculation, we may do that incorrectly today.

So close that up now to avoid the issue later. Also, avoid feature
flag checking at all if the internal format version is "2", which
did not have the flags, and also fail the check if the version is not
one of the two we know of (i.e. 2 or 3).

Change-Id: I352d6eceff2a90dd4a29d9eeeed1360f7b02dafa
Related-Bug: #2073413