Add GPT/MBR inspector
What most openstack services consider a "raw disk" is actually a GPT
(or MBR, which is a subset) layout. In order to distinguish between
an actual raw "unstructured series of bytes" from formats we can and
should recognize, this inspector should vastly reduce the number of
places we have to consider something as raw because it doesn't match
anything else we know about.

Change-Id: I1dac89f2118d845aa2e2ea74eab53017c8b3d09c