Add stream-based detection
This effectively unifies the InfoWrapper and detect_file_format()
behavior into a single approach that can work stream-based or with
a file on disk.

Note that the wrapper name is also changed as "InfoWrapper" was
named for glance's intended use-case, which was metadata extraction.
Since this has obviously grown in scope, now is the time to change
that name.

Change-Id: Id61f7472f791fc258ec7d0238568c379a7b27823