Avoid detecting FAT VBR as an MBR
The 1980s FAT filesystem has a VBR in the first sector, which looks
almost exactly like an MBR with zero partitions. To avoid detecting
these as MBRs, look for some extra attributes that indicate that the
structure is a VBR and avoid matching it as a GPT/MBR in that case.

We can add an inspector for this as a separate thing, but at the
moment we don't have that immediate need.

Closes-Bug: #2079850
Change-Id: Ibad87743b5a3b6469bd708d4caafe7911b045855