Squelch irrelevant format complaints
When we're eating a multi-inspector stream, we would expect the other
formats to fail at some point in the stream when they don't match.
This makes us not log the irrelevant format failures when we're
looking for a specific one. We were already logging that at debug
level for the sake of visibility into the process, but some people
have been misled by the complaint. In addition, soften the language
a bit to make it clear it's not an actual failure.

Change-Id: Ife973cead13eba0900b72d9234253d543f723636