Abort the stream early if no match
If we're reading a stream and expecting a given format, we should
abort as soon as we're sure that what we've read is not a match. In
the case of glance where the user has already promised to upload a
given format, we want to abort before we've read tons of traffic only
to reject the image later.

Related to blueprint glance-as-defender
Change-Id: Idf254b242d76ad0bfa9b06babc368d8b92f1c15f