Run mypy from tox
This avoids the need to duplicate our dependency list in multiple
places and allows us to take advantage of tox's dependency management
infrastructure, to ensure we always get the latest and greatest version
of a package allowed by upper-constraints.

While here, we also bump the other pre-commit hooks and update hints to
accommodate typed fixtures.

Change-Id: I4020374b45c8ead4b3b65651389c1e903a1de7bd
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>